refactor: restructure; add eslint and fix; new Buffer to Buffer.from
