+++ .eslintrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+    extends: 'standard',
+    globals: {
+        test: true,
+        expect: true,
+        describe: true,
+        jest: true
+    }
+};
