+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,6 +1,7 @@
+Please ensure all pull requests are made against the `develop` branch.
+
 *Issue #, if available:*
 
 *Description of changes:*
-
 
 By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.
