--- .jshintrc
@@ -1,11 +1,0 @@
-{
-  "esversion": 6,
-  "node": true,
-  "asi": true,
-  "globals": {
-    /* jest */
-    "test": false,
-    "expect": false,
-    "describe": false
-  }
-}
