+++ examples/basic-starter/.gitignore
