+++ examples/basic-starter/api-gateway-event.json
