+++ examples/basic-starter/app.local.js
