+++ examples/basic-starter/cloudformation.yaml
