+++ examples/basic-starter/lambda.js
@@ -1,5 +1,5 @@
 'use strict'
-const awsServerlessExpress = require(process.env.NODE_ENV === 'test' ? '../index' : 'aws-serverless-express')
+const awsServerlessExpress = require(process.env.NODE_ENV === 'test' ? '../src/index' : 'aws-serverless-express')
 const app = require('./app')
 
 // NOTE: If you get ERR_CONTENT_DECODING_FAILED in your browser, this is likely
