+++ examples/basic-starter/package.json
