+++ examples/basic-starter/sam-logo.png
