+++ examples/basic-starter/simple-proxy-api.yaml
